+++ src/pages/SignUp/components/SignUpForm.tsx
@@ -1,5 +1,5 @@
 import { RED_BUTTON_STYLE_AUTH } from "@/common/buttonStyles";
-import BasicButton from "@/common/components/atoms/BasicButton";
+import BasicButton from "@/common/components/atoms/button/BasicButton";
 import FormInputField from "@/common/components/molecules/FormInputField";
 import { AuthFormType } from "@/common/types/formTypes";
 import { validationRules } from "@/common/validationRules";
