+++ src/pages/SignUp/index.tsx
@@ -1,6 +1,6 @@
 import SignUpForm from "@/pages/SignUp/components/SignUpForm";
 import AuthFirstHeading from "@/common/components/atoms/AuthFirstHeading";
-import LinkButton from "@/common/components/atoms/LinkButton";
+import LinkButton from "@/common/components/atoms/button/LinkButton";
 import { RED_TEXT_STYLE_AUTH } from "@/common/buttonStyles";
 
 export default function SignUp() {
